+++ features/record/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './use-distance-page-modal';
 export * from './use-get-browser-width';
 export * from './use-pool-search-page-modal';
 export * from './use-sub-info-text-fields';
+export * from './use-time-bottom-sheet';
